+++ migrations/env.py
@@ -32,6 +32,7 @@ from app.models.allegro_token import AllegroToken
 from app.models.allegro_order import AllegroOrder, AllegroLineItem, AllegroBuyer, OrderLineItem
 from app.models.warehouse import Product, Stock, Sale, Transfer
 from app.models.operations import Operation
+from app.models.allegro_event_tracker import AllegroEventTracker
 
 target_metadata = SQLModel.metadata